Split readtoarray.c into add_to_list.c ; check_valid.c ; readfromfile.c to cope with norm. Added some test-files. Fixed norm on most of files.
