+++ Makefile
@@ -6,14 +6,17 @@
 #    By: marius <marius@student.42.fr>              +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/01/11 14:35:22 by marius            #+#    #+#              #
-#    Updated: 2022/01/17 12:44:52 by severi           ###   ########.fr        #
+#    Updated: 2022/01/19 01:44:24 by severi           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
 NAME = fillit
 SRCS = includes/main.c \
 		includes/readtoarray.c \
-		includes/solve.c 
+		includes/solve.c \
+		includes/readfromfile.c \
+		includes/add_to_list.c \
+		includes/check_valid.c 
 OBJS = includes/*.o
 HDR = includes/fillit.h
 LHDR = libft/
@@ -38,6 +41,6 @@ fclean: clean
 
 re: fclean all
 
-debug: re 
+debug: fclean
 	make -C $(LIBFT)
 	gcc $(FLAGS) $(DFLAGS) -o $(NAME) $(SRCS)-I $(LHDR) $(HDR) -L. libft/libft.a
