+++ eval_tests/bad_tetrimino_2nd_empty
@@ -1,0 +1,6 @@
+....
+....
+....
+####
+
+....
