+++ eval_tests/bad_tetrimino_5_chars
@@ -1,0 +1,4 @@
+....
+.##.
+.###
+....
