+++ eval_tests/bad_tetrimino_5x5square
@@ -1,0 +1,5 @@
+.....
+...##
+...##
+.....
+.....
