+++ eval_tests/bad_tetrimino_empty
@@ -1,0 +1,4 @@
+....
+....
+....
+....
