+++ eval_tests/bad_tetrimino_last_with_comma
@@ -1,0 +1,14 @@
+....
+.##.
+.##.
+....
+
+....
+.##.
+.##.
+....
+
+....
+.##.
+.##.
+..,.
