+++ eval_tests/bad_tetrimino_nl_at_EOF
@@ -1,0 +1,5 @@
+....
+.##.
+.##.
+....
+
