+++ eval_tests/bad_tetrimino_trimino
@@ -1,0 +1,4 @@
+....
+.##.
+..#.
+....
