+++ includes/solve.c
@@ -6,13 +6,12 @@
 /*   By: severi <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/17 11:04:48 by severi            #+#    #+#             */
-/*   Updated: 2022/01/17 11:06:16 by severi           ###   ########.fr       */
+/*   Updated: 2022/01/19 01:15:46 by severi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fillit.h"
 
-void	solve()
+void	solve(void)
 {
-
 }
